Will use pdmp for low SN obs and now creates a results json
