+++ nextflow.config
@@ -13,7 +13,6 @@ params.pulsar   = ""    // Pulsar name for PSRDB search - returns only observati
 
 // Processing options
 params.use_edge_subints = false // Use first and last 8 second subints of observation archives
-params.fluxcal = false // Calibrate flux densities. Should only be done for calibrator observations
 params.tos_sn = 12 // Desired TOA S/N ratio, used to calculate the nsub to use
 params.nchans = [1, 4, 16] // List of nchans to frequency scrunch the data into
 params.npols = [1] // List of number of polarisations to scrunch the data into
@@ -130,8 +129,8 @@ if ( hostname.startsWith("tooarrana") ) {
             // queue = 'workq'
             cpus = 1
             executor = 'slurm'
-            errorStrategy = 'retry'
-            maxRetries = 2
+            // errorStrategy = 'retry'
+            // maxRetries = 2
         }
         // Software dependency set up
         withLabel: meerpipe {
