+++ pyproject.toml
@@ -13,7 +13,7 @@ python = "^3.10"
 
 [tool.poetry.scripts]
 fluxcal = "meerpipe.scripts.fluxcal:main"
-generate_images = "meerpipe.scripts.generate_images:main"
+generate_images_results = "meerpipe.scripts.generate_images_results:main"
 
 [build-system]
 requires = ["poetry-core>=1.0.0"]
